Brings the machine-independent LP64 sweep (FrameAllocator/Heap/memory_manager/
Console/Framebuffer/String/List/BlockDevice/mmapInfo/off_t+LinuxStat/kernelSyscall
ABI) onto the x86_64 foundation alongside Plans 3-4 (paging + interrupts).
Verified: i686 build+check-arch, host tests i686 606 / x86_64 613, staged x86_64
QEMU boot (banner + PIT 466 ticks + zero faults).

+++ arch/x86/drivers/AtaBlockDevice.cpp
@@ -11,7 +11,9 @@ AtaBlockDevice::AtaBlockDevice(const char* name) {
 // once per command and then transfers count*256 words; for count > 1 that
 // desyncs the drive (the data buffer must be re-serviced per sector), which
 // corrupts this and subsequent reads. One sector per command keeps it correct.
-int AtaBlockDevice::readSectors(unsigned lba, unsigned count, void* buf) {
+// lba is 64-bit at the HAL, but the ATA.S PIO path is LBA28 (Hdd takes an int): the (int)
+// cast is a conscious 28-bit truncation — fine for our QEMU disk; LBA48 is future work.
+int AtaBlockDevice::readSectors(uint64_t lba, unsigned count, void* buf) {
 	char* p = (char*) buf;
 	for (unsigned i = 0; i < count; i++)
 		if (Hdd::readSectors((int) (lba + i), 1, p + i * 512) != 0)
@@ -19,7 +21,7 @@ int AtaBlockDevice::readSectors(unsigned lba, unsigned count, void* buf) {
 	return 0;
 }
 
-int AtaBlockDevice::writeSectors(unsigned lba, unsigned count, const void* buf) {
+int AtaBlockDevice::writeSectors(uint64_t lba, unsigned count, const void* buf) {
 	char* p = (char*) buf;
 	for (unsigned i = 0; i < count; i++)
 		if (Hdd::writeSectors((int) (lba + i), 1, p + i * 512) != 0)
